+++ cart-rest-api-for-woocommerce.php
@@ -2,10 +2,10 @@
 /*
  * Plugin Name: CoCart
  * Plugin URI:  https://cocart.xyz
- * Description: A REST-API for WooCommerce that enables the ability to add, view, update and delete items from the cart.
+ * Description: Provides additional REST-API endpoints for WooCommerce to enable the ability to add, view, update and delete items from the cart.
  * Author:      Sébastien Dumont
  * Author URI:  https://sebastiendumont.com
- * Version:     1.1.2
+ * Version:     1.2.0
  * Text Domain: cart-rest-api-for-woocommerce
  * Domain Path: /languages/
  *
@@ -18,11 +18,11 @@
  * License URI: http://www.gnu.org/licenses/gpl-3.0.html
  */
 
-if ( ! class_exists( 'WC_Cart_Endpoint_REST_API' ) ) {
-	class WC_Cart_Endpoint_REST_API {
+if ( ! class_exists( 'CoCart' ) ) {
+	class CoCart {
 
 		/**
-		 * @var WC_Cart_Endpoint_REST_API - the single instance of the class.
+		 * @var CoCart - the single instance of the class.
 		 *
 		 * @access protected
 		 * @static
@@ -37,7 +37,7 @@ if ( ! class_exists( 'WC_Cart_Endpoint_REST_API' ) ) {
 		 * @static
 		 * @since  1.0.0
 		 */
-		public static $version = '1.1.2';
+		public static $version = '1.2.0';
 
 		/**
 		 * Required WooCommerce Version
@@ -49,15 +49,16 @@ if ( ! class_exists( 'WC_Cart_Endpoint_REST_API' ) ) {
 		public static $required_woo = '3.0.0';
 
 		/**
-		 * Main WC_Cart_Endpoint_REST_API Instance.
+		 * Main CoCart Instance.
 		 *
-		 * Ensures only one instance of WC_Cart_Endpoint_REST_API is loaded or can be loaded.
+		 * Ensures only one instance of CoCart is loaded or can be loaded.
 		 *
-		 * @access public
+		 * @access  public
 		 * @static
-		 * @since  1.0.0
-		 * @see    WC_Cart_Endpoint_REST_API()
-		 * @return WC_Cart_Endpoint_REST_API - Main instance
+		 * @since   1.0.0
+		 * @version 1.2.0
+		 * @see     CoCart()
+		 * @return  CoCart - Main instance
 		 */
 		public static function instance() {
 			if ( is_null( self::$_instance ) ) {
@@ -91,60 +92,107 @@ if ( ! class_exists( 'WC_Cart_Endpoint_REST_API' ) ) {
 		/**
 		 * Load the plugin.
 		 *
-		 * @access public
-		 * @since  1.0.0
+		 * @access  public
+		 * @since   1.0.0
+		 * @version 1.2.0
 		 */
 		public function __construct() {
-			// Check WooCommerce dependency.
-			add_action( 'plugins_loaded', array( $this, 'check_woocommerce_dependency' ) );
+			$this->setup_constants();
+			$this->includes();
 
-			// Add link to documentation and support.
-			add_filter( 'plugin_row_meta', array( $this, 'plugin_row_meta' ), 10, 2 );
-
-			// Include required files
-			add_action( 'init', array( $this, 'includes' ) );
+			// Initialize plugin.
+			add_action( 'plugins_loaded', array( $this, 'initialize_plugin' ) );
 
 			// Load translation files.
 			add_action( 'init', array( $this, 'load_plugin_textdomain' ) );
-		}
+		} // END __construct()
 
 		/**
-		 * Get the Plugin Path.
+		 * Setup Constants
 		 *
-		 * @access public
-		 * @static
-		 * @since  1.0.0
-		 * @return string
+		 * @access private
+		 * @since  1.2.0
 		 */
-		public static function plugin_path() {
-			return untrailingslashit( plugin_dir_path( __FILE__ ) );
-		} // END plugin_path()
+		private function setup_constants() {
+			$this->define('COCART_VERSION', self::$version);
+			$this->define('COCART_FILE', __FILE__);
+			$this->define('COCART_SLUG', 'cart-rest-api-for-woocommerce');
+
+			$this->define('COCART_URL_PATH', untrailingslashit( plugins_url( '/', __FILE__ ) ) );
+			$this->define('COCART_FILE_PATH', untrailingslashit( plugin_dir_path( __FILE__ ) ) );
+
+			$this->define('COCART_WP_VERSION_REQUIRE', '4.4');
+
+			$this->define('COCART_STORE_URL', 'https://cocart.xyz/');
+			$this->define('COCART_PLUGIN_URL', 'https://wordpress.org/plugins/cart-rest-api-for-woocommerce/');
+			$this->define('COCART_SUPPORT_URL', 'https://wordpress.org/support/plugin/cart-rest-api-for-woocommerce');
+			$this->define('COCART_REVIEW_URL', 'https://wordpress.org/support/plugin/cart-rest-api-for-woocommerce/reviews/');
+			$this->define('COCART_DOCUMENTATION_URL', 'https://co-cart.github.io/co-cart-docs/');
+		} // END setup_constants()
 
 		/**
-		 * Check WooCommerce dependency before activation.
+		 * Define constant if not already set.
+		 *
+		 * @access private
+		 * @since  1.2.0
+		 * @param  string $name
+		 * @param  string|bool $value
+		 */
+		private function define( $name, $value ) {
+			if ( ! defined( $name ) ) {
+				define( $name, $value );
+			}
+		} // END define()
+
+		/**
+		 * Initialize plugin
 		 *
 		 * @access public
-		 * @since  1.0.0
 		 */
-		public function check_woocommerce_dependency() {
-			// Check we're running the required version of WooCommerce.
-			if ( ! defined( 'WC_VERSION' ) || version_compare( WC_VERSION, self::$required_woo, '<' ) ) {
-				add_action( 'admin_notices', array( $this, 'admin_notice' ) );
+		public function initialize_plugin() {
+			// If the current user can not install plugins then return nothing!
+			if ( ! current_user_can( 'install_plugins' ) ) {
 				return false;
 			}
-		} // END check_woocommerce_dependency()
+
+			// Check we're running the required version of WooCommerce.
+			if ( ! defined( 'WC_VERSION' ) || version_compare( WC_VERSION, self::$required_woo, '<' ) ) {
+				add_action( 'admin_notices', array( $this, 'requirement_wc_notice' ) );
+				return false;
+			}
+		} // END initialize_plugin()
 
 		/**
 		 * Display a warning message if minimum version of WooCommerce check fails.
 		 *
 		 * @access  public
 		 * @since   1.0.0
-		 * @version 1.1.2
+		 * @version 1.2.0
 		 * @return  void
 		 */
-		public function admin_notice() {
-			echo '<div class="notice notice-error"><p>' . sprintf( __( '%1$s requires at least %2$s v%3$s or higher.', 'cart-rest-api-for-woocommerce' ), esc_html__( 'CoCart', 'cart-rest-api-for-woocommerce' ), esc_html__( 'WooCommerce', 'cart-rest-api-for-woocommerce' ), self::$required_woo ) . '</p></div>';
-		} // END admin_notice()
+		public function requirement_wc_notice() {
+			include( dirname( __FILE__ ) . '/views/html-notice-requirement-wc.php' );
+		} // END requirement_wc_notice()
+
+		/**
+		 * Includes Cart REST-API for WooCommerce.
+		 *
+		 * @access  public
+		 * @since   1.0.0
+		 * @version 1.2.0
+		 * @return  void
+		 */
+		public function includes() {
+			include_once( dirname( __FILE__ ) . '/includes/class-wc-cart-rest-api-init.php' );
+
+			// Include admin class to handle all back-end functions.
+			if ( is_admin() ) {
+				include_once( dirname( __FILE__ ) . '/includes/admin/class-cocart-admin.php' );
+			}
+
+			// Install.
+			require_once( dirname( __FILE__ ) . '/includes/class-cocart-install.php' );
+		} // END includes()
 
 		/**
 		 * Make the plugin translation ready.
@@ -160,44 +208,8 @@ if ( ! class_exists( 'WC_Cart_Endpoint_REST_API' ) ) {
 			load_plugin_textdomain( 'cart-rest-api-for-woocommerce', false, dirname( plugin_basename( __FILE__ ) ) . '/languages/' );
 		} // END load_plugin_textdomain()
 
-		/**
-		 * Includes Cart REST API for WooCommerce Admin.
-		 *
-		 * @access public
-		 * @since  1.0.0
-		 * @return void
-		 */
-		public function includes() {
-			include_once( $this->plugin_path() . '/includes/class-wc-cart-rest-api-init.php' );
-		} // END includes()
-
-		/**
-		 * Show row meta on the plugin screen.
-		 *
-		 * @access  public
-		 * @static
-		 * @since   1.0.0
-		 * @version 1.0.9
-		 * @param   mixed $links
-		 * @param   mixed $file
-		 * @return  array
-		 */
-		public static function plugin_row_meta( $links, $file ) {
-			if ( $file == plugin_basename( __FILE__ ) ) {
-				$row_meta = array(
-					'docs'    => '<a href="' . esc_url( 'https://co-cart.github.io/co-cart-docs/' ) . '" target="_blank" aria-label="' . esc_attr( __( 'View CoCart Documentation', 'cart-rest-api-for-woocommerce' ) ) . '">' . __( 'Documentation', 'cart-rest-api-for-woocommerce' ) . '</a>',
-					'support' => '<a href="' . esc_url( 'https://wordpress.org/support/plugin/cart-rest-api-for-woocommerce/' ) . '" target="_blank" aria-label="' . esc_attr( __( 'Create a support ticket for CoCart', 'cart-rest-api-for-woocommerce' ) ) . '">' . __( 'Support', 'cart-rest-api-for-woocommerce' ) . '</a>',
-					'review'  => '<a href="' . esc_url( 'https://wordpress.org/plugins/cart-rest-api-for-woocommerce/#reviews' ) . '" target="_blank" aria-label="' . esc_attr( __( 'Review CoCart on WordPress.org', 'cart-rest-api-for-woocommerce' ) ) . '">' . __( 'Leave a Review', 'cart-rest-api-for-woocommerce' ) . '</a>',
-				);
-
-				$links = array_merge( $links, $row_meta );
-			}
-
-			return $links;
-		} // END plugin_row_meta()
-
 	} // END class
 
 } // END if class exists
 
-return WC_Cart_Endpoint_REST_API::instance();
+return CoCart::instance();